Move DefaultSerializer into its own file and rename
