+++ src/common/serializer/index.ts
@@ -1,1 +1,2 @@
 export * from './serializers';
+export * from './DefaultSerializer';
